+++ plugin-build/src/main/kotlin/io/sentry/android/gradle/util/SentryPluginUtils.kt
@@ -1,9 +1,12 @@
 package io.sentry.android.gradle.util
 
+import com.android.build.gradle.api.ApplicationVariant
 import java.util.Locale
+import org.gradle.api.Project
 import org.gradle.api.Task
 import org.gradle.api.logging.Logger
 import org.gradle.api.tasks.TaskProvider
+import proguard.gradle.plugin.android.dsl.ProGuardAndroidExtension
 
 internal object SentryPluginUtils {
 
@@ -19,5 +22,17 @@ internal object SentryPluginUtils {
         ""
     } else {
         substring(0, 1).toUpperCase(Locale.US) + substring(1)
+    }
+
+    fun isMinificationEnabled(project: Project, variant: ApplicationVariant): Boolean {
+        var isConfiguredWithGuardsquareProguard = false
+        project.plugins.withId("com.guardsquare.proguard") {
+            val proguardExtension = project.extensions.getByType(
+                ProGuardAndroidExtension::class.java
+            )
+            val variantConfiguration = proguardExtension.configurations.findByName(variant.name)
+            isConfiguredWithGuardsquareProguard = variantConfiguration != null
+        }
+        return isConfiguredWithGuardsquareProguard || variant.buildType.isMinifyEnabled
     }
 }